Lots of changes on this commit
moved login and register forms into separate components for easier error handling

created an ErrPopOver component to handle all Popovers had the same styles

removed import Popover and popover consts from components now using ErrPopover.jsx

deleted console log from Home.js

api/index.js deleted unused vars and all the console.logs
