+++ src/components/EditRoutineModal.jsx
@@ -1,11 +1,11 @@
-import { useEffect, useState } from 'react'
+import { useEffect, useState, createRef } from 'react'
 import Modal from 'react-bootstrap/Modal'
 import Form from 'react-bootstrap/Form'
 import Button from 'react-bootstrap/Button'
-import Popover from 'react-bootstrap/Popover'
 import Container from 'react-bootstrap/Container'
 import OverlayTrigger from 'react-bootstrap/OverlayTrigger'
 import { editRoutine, fetchRoutines } from '../api'
+import ErrPopover from './ErrPopover'
 
 export default function EditRoutineModal({
   show,
@@ -18,6 +18,7 @@ export default function EditRoutineModal({
   const [goal, setGoal] = useState('')
   const [errMessage, setErrMessage] = useState('')
   const [errShow, setErrShow] = useState('')
+  const ref = createRef()
 
   useEffect(() => {
     setGoal(routineGoal)
@@ -48,15 +49,6 @@ export default function EditRoutineModal({
     setErrShow(false)
     setErrMessage('')
   }
-
-  const popover = (
-    <Popover id='popover-basic' className='bg-danger'>
-      <Popover.Header as='h3' className='bg-danger text-light'>
-        {errMessage.name}
-      </Popover.Header>
-      <Popover.Body className='text-light'>{errMessage.message}</Popover.Body>
-    </Popover>
-  )
 
   useEffect(() => {
     const timerId = setInterval(() => {
@@ -101,7 +93,11 @@ export default function EditRoutineModal({
             />
           </Form.Group>
           <Container className='d-flex justify-content-end gap-2 w-100'>
-            <OverlayTrigger show={errShow} overlay={popover} placement='bottom'>
+            <OverlayTrigger
+              show={errShow}
+              overlay={<ErrPopover ref={ref} errmessage={errMessage} />}
+              placement='bottom'
+            >
               <Button variant='success' type='submit'>
                 Update
               </Button>
